Refactors code to remove some code duplication
