+++ YuzuDelivery.Core/HbsHelpers/ModPartial.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using HandlebarsDotNet.Compiler;
 using Newtonsoft.Json;
+using YuzuDelivery.Core.Helpers;
 
 namespace YuzuDelivery.Core
 {
@@ -54,7 +55,7 @@ namespace YuzuDelivery.Core
                             var test = "";
                         }
 
-                        r(writer, GetDataModel(parameters) ?? parameters[1]);
+                        r(writer, PartialHelpers.GetDataModel(parameters) ?? parameters[1]);
                     }
                     else
                         throw new Exception(string.Format("Handlebars modifier partial cannot find partial {0}",
@@ -65,55 +66,8 @@ namespace YuzuDelivery.Core
                         "Handlebars modifier partial should have 3 parameters; parial name, content and modifier");
             });
         }
-        private static bool IsSimple(Type type)
-        {
-            return type.IsPrimitive
-                   || type == typeof(string);
-        }
 
-        private static Dictionary<string, object> GetDataModel(object[] parameters)
-        {
-            var paramType = parameters[1].GetType();
-            //we can't support modifiers and hashParameters on base types
-            if (IsSimple(paramType))
-            {
-                return null;
-            }
-            
-            var modifiers = parameters.Where((source, index) => index > 1)
-                .Where(x => x != null && x.GetType() != typeof(HashParameterDictionary))
-                .Select(x => x.ToString()).ToList();
-            
-            //not sure what is faster
-            // linq:
-            var properties = parameters[1].GetType().GetProperties().ToDictionary(
-              property => StringExtensions.FirstCharacterToLower(property.Name),
-              property => property.GetValue(parameters[1]));
-
-            //jsonSerialize/deserialize:
-            //var json = JsonConvert.SerializeObject(parameters[1]);
-            //var properties = JsonConvert.DeserializeObject<Dictionary<string, object>>(json); 
-
-            if (properties.Any(property => property.Key == "_modifiers" && property.Value == null))
-            {
-                properties.Remove("_modifiers");
-            }
-
-            if (!properties.ContainsKey("_modifiers"))
-            {
-                properties.Add("_modifiers", modifiers);
-            }
-
-            if (!(parameters[parameters.Length - 1] is HashParameterDictionary hashParameterDictionary))
-                return properties;
-
-            foreach (var parameter in hashParameterDictionary)
-            {
-                properties.Add(parameter.Key, parameter.Value);
-            }
-
-            return properties;
-        }
+        
     }
